Fixing moving screen down. Logs in different thread

+++ Going_UP/Structures.h
@@ -25,5 +25,5 @@ struct GameLevel
 	int maxX;// = GAME_WIDTH;
 	int maxY;// = GAME_HEIGHT * 5;
 	int curX;
-	int curY;
+	double curY;
 };
